+++ include/clap/ext/event-registry.h
@@ -14,7 +14,7 @@ typedef struct clap_host_event_registry {
    //
    // Return false and sets *space to UINT16_MAX if the space name is unknown to the host.
    // [main-thread]
-   bool (*query)(const clap_host_t *host, const char *space_name, uint16_t *space_id);
+   bool(CLAP_ABI *query)(const clap_host_t *host, const char *space_name, uint16_t *space_id);
 } clap_host_event_registry_t;
 
 #ifdef __cplusplus