+++ include/clap/ext/note-ports.h
@@ -42,14 +42,14 @@ typedef struct clap_note_port_info {
 typedef struct clap_plugin_note_ports {
    // number of ports, for either input or output
    // [main-thread]
-   uint32_t (*count)(const clap_plugin_t *plugin, bool is_input);
+   uint32_t(CLAP_ABI *count)(const clap_plugin_t *plugin, bool is_input);
 
    // get info about about a note port.
    // [main-thread]
-   bool (*get)(const clap_plugin_t   *plugin,
-               uint32_t               index,
-               bool                   is_input,
-               clap_note_port_info_t *info);
+   bool(CLAP_ABI *get)(const clap_plugin_t   *plugin,
+                       uint32_t               index,
+                       bool                   is_input,
+                       clap_note_port_info_t *info);
 } clap_plugin_note_ports_t;
 
 enum {
@@ -66,11 +66,11 @@ enum {
 typedef struct clap_host_note_ports {
    // Query which dialects the host supports
    // [main-thread]
-   uint32_t (*supported_dialects)(const clap_host_t *host);
+   uint32_t(CLAP_ABI *supported_dialects)(const clap_host_t *host);
 
    // Rescan the full list of note ports according to the flags.
    // [main-thread]
-   void (*rescan)(const clap_host_t *host, uint32_t flags);
+   void(CLAP_ABI *rescan)(const clap_host_t *host, uint32_t flags);
 } clap_host_note_ports_t;
 
 #ifdef __cplusplus