+++ include/clap/ext/timer-support.h
@@ -10,7 +10,7 @@ extern "C" {
 
 typedef struct clap_plugin_timer_support {
    // [main-thread]
-   void (*on_timer)(const clap_plugin_t *plugin, clap_id timer_id);
+   void(CLAP_ABI *on_timer)(const clap_plugin_t *plugin, clap_id timer_id);
 } clap_plugin_timer_support_t;
 
 typedef struct clap_host_timer_support {
@@ -18,10 +18,10 @@ typedef struct clap_host_timer_support {
    // The host may adjust the period if it is under a certain threshold.
    // 30 Hz should be allowed.
    // [main-thread]
-   bool (*register_timer)(const clap_host_t *host, uint32_t period_ms, clap_id *timer_id);
+   bool(CLAP_ABI *register_timer)(const clap_host_t *host, uint32_t period_ms, clap_id *timer_id);
 
    // [main-thread]
-   bool (*unregister_timer)(const clap_host_t *host, clap_id timer_id);
+   bool(CLAP_ABI *unregister_timer)(const clap_host_t *host, clap_id timer_id);
 } clap_host_timer_support_t;
 
 #ifdef __cplusplus